Fix guassian fit of parameter distributions
Previously, the normalization of the gaussian fit was a free
parameter, because it is difficult to normalize properly when
part of the distribution is cut off. However, this caused the
fit to fail in many cases.

This commit forces the gaussian to be normalized to 1, and
renormalizes the histogram so it is normalized to 1.